feat(autocmds): Add last location jump on buffer open
This commit introduces a new autocmd for BufReadPost event. The purpose
of this autocmd is to return the cursor to the last location when a
buffer is opened. This feature is useful for navigating large files and
improves user experience. The autocmd excludes certain filetypes like
'gitcommit', 'commit', 'gitrebase' and buffers that have already
triggered the autocmd.